// src/types/WorkEntry.ts
export interface WorkEntry {
  date: string;
  fullName: string;
  entryType: string;
  time: string;
  duration: string;
  break: string;
  breakType: string;
  client: string;
  projectCode: string;
  project: string;
  notes: string;
  createdOn: string;
  lastEditedBy: string;
  lastEditedOn: string;
}